http-api: Move DefaultHttpLoadBalancerFactory from http-netty to http-api
Motivation:

There is nothing netty specific about DefaultHttpLoadBalancerFactory
but it lives in the servicetalk-http-netty package.

Modifications:

- Copy the class to the http-api package.
- Deprecate the http-netty version.

Result:

What is the result of this change?
